Luck-based FindRandomPost implementation
By having $sample as the first pipeline operator, we use the pseudo-random behavior as documented in https://docs.mongodb.com/manual/reference/operator/aggregation/sample/
which has far better performances for big databases.
The drawback is the off-chance that we may end up with an empty result set. For this reason, we added 2 additional iterations, this should be an effective tradeoff for document sets with a limited number of "bad" documents